+++ source/vector_operations.h
@@ -16,11 +16,11 @@ std::vector<double> &operator-=(std::vector<double> &vec1,
 std::vector<double> operator-(const std::vector<double> &vec1,
                               const std::vector<double> &vec2);
 
-std::vector<double> &operator*=(std::vector<double> &vec, const int num);
+std::vector<double> &operator*=(std::vector<double> &vec, const double num);
 
-std::vector<double> operator*(const std::vector<double> &vec, const int num);
+std::vector<double> operator*(const std::vector<double> &vec, const double num);
 
-std::vector<double> operator*(const int num, const std::vector<double> &vec);
+std::vector<double> operator*(const double num, const std::vector<double> &vec);
 
 double dot_product(const std::vector<double> &vec1,
                    const std::vector<double> &vec2);
